Create Binary_Search_Algorithm
Binary search algorithms are also known as half interval search. They return the position of a target value in a sorted list.
These algorithms use the “divide and conquer” technique to find the value's position.
Binary search algorithms and linear search algorithms are examples of simple search algorithms.
In binary search, the middle element in the list is found before comparing with the key value you are searching for. But in linear search, the elements are taken one by one in the list by looping through and comparing with the key value.