MEMORY PUZZLE GAME
In this game we will create 3 levels: easy, medium and hard. At the start of the game the shapes are hidden. The player clicks on the tiles to uncover the tile and when two similar tiles are uncovered the score increases. To complete the game the user has to uncover all the pairs of shapes. After completion of the game, the number of moves required to complete the game is also displayed on the screen.  The main objective of the project is to develop a Memory Puzzle Game that has 3 different levels. Installation of tkinter is a must to start developing the project. Project File Structure
1. Installing Tkinter
2. Importing Modules and initializing tkinter window
3. Various functions required for creating easy level
4. Various functions required for creating medium level
5. Various functions required for creating hard level
1> Installing Tkinter
Installation of tkinter is a must to start the project. Tkinter is a module that provides the most easy way to develop a graphical user interface. To install tkinter write the command mentioned below on your terminal window or cmd.
pip install tkinter
2>Importing Modules and initializing tkinter window
from tkinter import *
import random
from tkinter import ttk
PuzzleWindow=Tk()
PuzzleWindow.title('Memory Puzzle Game By DataFlair')
tabs = ttk.Notebook(PuzzleWindow) 
easy= ttk.Frame(tabs)
3> Various functions required for creating easy level
4> Various function required for creating medium level
5> Various function required for creating hard level